day-08/requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.7
python-dotenv>=1.0.0